requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
PyMySQL==1.0.2
python-dotenv==0.19.2
mysql-connector==2.2.9
mysql-connector-python==9.3.0
google-genai==1.21.1
pandas==2.3.0
psycopg2-binary==2.9.9
werkzeug==2.0.3
SQLAlchemy==1.4.29
cryptography==45.0.4
gunicorn>=21.2.0
PyMySQL==1.0.2
jwt==1.4.0
PyJWT==2.10.1
flask-cors==6.0.1